requirements.txt
werkzeug==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Faker==26.0.0
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
SQLAlchemy==2.0.31
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
redis
celery
flower
gunicorn
wheel
zipp>=3.20
ffmpeg
moviepy
pytest
pytest-flask
pytest-cov
psycopg2-binary
google-cloud-storage
